DEJA VU
[Intro: Instrumental]

[Verse 1]
See your body into the moonlight
Even if I try to cancel
All the pictures into the mind
There's a flashing in my eyes

[Hook]
Don't you see my condition?
The fiction - is gonna run it again
Can't you see now illusions?
Right into your mind

[Chorus]
Deja vu
I've just been in this place before
Higher on the street
And I know it's my time to go
Calling you, and the search is a mystery
Standing on my feet
It's so hard when I try to be me, woah

Deja vu
I've just been in this time before
Higher on the beat
And I know it's a place to go
Calling you and the search is a mystery
Standing on my feet
It's so hard when I try to be me, yeah

[Verse 2]
See the future into the present
See my past leaves in the distance
Try to guess now what's going on
And the band begins to play

[Hook]
Don't you see my condition?
The fiction - is gonna run it again
Can't you see now illusions?
Right into your mind

[Chorus]
Deja vu
I've just been in this place before
Higher on the street
And I know it's my time to go
Calling you, and the search is a mystery
Standing on my feet
It's so hard when I try to be me, woah

Deja vu
I've just been in this time before
Higher on the beat
And I know it's a place to go
Calling you and the search is a mystery
Standing on my feet
It's so hard when I try to be me, yeah

[Verse 1]
See your body into the moonlight
Even if I try to cancel
All the pictures into the mind
There's a flashing in my eyes

[Hook]
Don’t you see my condition, the fiction
Has gone running again
Can't you see now, illusions
Right into your mind

[Chorus]
Deja vu
I've just been in this place before
Higher on the street
And I know it's my time to go
Calling you, and the search is a mystery
Standing on my feet
It's so hard when I try to be me, woah

Deja vu
I've just been in this time before
Higher on the beat
And I know it's a place to go
Calling you and the search is a mystery
Standing on my feet
It's so hard when I try to be me, yeah